- converted a lot of "void blah();" to "void blah( void );" fo avoid a silly warning in vs10 - converted a lot of "FACING_T blah; int flah; blah += flah;" to "FACING_T blah; int flah; blah = (int)blah + flah;" - added a BOOL_T macro to convert c++ boolean type to bool_t, and implemented it a couple of places. - cahnged the scancode_to_ascii arrays to use int because a lot of text functions return int. also added a warning to the function that generates these arrays to catch wierd values. - minor mods to some variable types - changed some of the optimization settings in vs10. It will now intercept floating point exceptions if you set your debugger to catch them. This is a debugging feature that will likely be turned off in the future.

+++ egolib/extensions/SDL_GL_extensions.c
@@ -44,10 +44,10 @@ static FILE * _SDL_GL_stdout = NULL;
 // will have to be changed to reflect that
 #if (SDL_BYTEORDER == SDL_LIL_ENDIAN)
 
-const Uint32 sdl_a_shift = 24;
-const Uint32 sdl_b_shift = 16;
-const Uint32 sdl_g_shift =  8;
-const Uint32 sdl_r_shift =  0;
+const Uint8 sdl_a_shift = 24;
+const Uint8 sdl_b_shift = 16;
+const Uint8 sdl_g_shift =  8;
+const Uint8 sdl_r_shift =  0;
 
 const Uint32 sdl_a_mask = ( Uint32 )( 0xFF << 24 );
 const Uint32 sdl_b_mask = ( Uint32 )( 0xFF << 16 );
@@ -56,10 +56,10 @@ const Uint32 sdl_r_mask = ( Uint32 )( 0xFF <<  0 );
 
 #else
 
-const Uint32 sdl_a_shift = 0;
-const Uint32 sdl_b_shift = 8;
-const Uint32 sdl_g_shift = 16;
-const Uint32 sdl_r_shift = 24;
+const Uint8 sdl_a_shift = 0;
+const Uint8 sdl_b_shift = 8;
+const Uint8 sdl_g_shift = 16;
+const Uint8 sdl_r_shift = 24;
 
 const Uint32 sdl_a_mask = ( Uint32 )( 0xFF <<  0 );
 const Uint32 sdl_b_mask = ( Uint32 )( 0xFF <<  8 );
